Allow spark dependency to be configured dynamically
Signed-off-by: Ahmed Hussein <ahussein@nvidia.com>

Fixes #1316

Allow user-tools to pick the SPARK dependencies based on a runtime
env_var. The value format follows the same format of `buildver` in the
scala pom file.
Currently 333 and 350 (default) are supported.
If user specifies an invalid value, there will be a warning message,
then the process fails running the java cmd.

**Changes**

- Add dependency key to the platform config-file
- A platform can define its own default dependency versions using
  `activeBuildVer` key
- Add a default `RUNTIME_BUILDVER` in the `__init__.py` to allow
  upgrades of spark release during official releases
- Read an env_var `RAPIDS_USER_TOOLS_RUNTIME_BUILDVER` to pick the
  correct dependency.
- Currently, only `333` and `350` are supported. Default is `350`