Allow users to specify custom Dependency jars
Signed-off-by: Ahmed Hussein <ahussein@nvidia.com>

Fixes #1359

Add a new input argument that takes a path to a yaml file `--tools_config_file`
The config file allows the users to define their own binaries that need
to be added to the classpath of the tools jar cmd.
This change is important because users can use the user-tools wrapper
with their custom spark.