Refactor Exec Parsers - unify the code calling similar methods
This PR is to remove redundant code in many ExecParsers. We have
several ExecParsers with the same functionality in parse code.
In this PR, following changes are done:

1. If the Execs don't have expressions, then execName is assigned and
ExecInfo object is created from GenericExecParser.
2. If the Execs have expressions then reflection is used to assign the
appropriate parser to be used.
3. If there is any additional calculation in the parse function, then
only that part is overriden in the ExecParser and the rest of the
function uses the GenereicExecParser base code
4.Removed all the files which are not needed anymore

Signed-off-by: Niranjan Artal <nartal@nvidia.com>