Report all operators in the output file
This PR is to print all the operators per app and per sqlID in a new
file.
This helps to get the count of operators in an application.
It has count of both supported and unsupported operators.
Added arguments in ExecInfo to store references of execName and
expressionNames.
ExecInfo analysis is done at the end to save memory overhead of storing
the information.
Fixed a bug where the parseAggregateExpressions was

Signed-off-by: Niranjan Artal <nartal@nvidia.com>