Use a stub to store Spark StageInfo (#1525)
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Fixes #1524

This commit uses a smaller class `StageInfoStub` to store Spark's
StageInfo. This class is common between all the spark implementations
but it has more fields to the constructor across different versions.

Currently we only use a subset of the class fields. The remaining fields
represent an overhead or redundant storage; especially when it comes to
store the accumulables and taskMetrics for each stage.

To evaluate the memory optimization, a new `Checkpoint` mechanism was
added to allow gathering information at separate stages of the
execution.
The `checkpoint` design and implementation can be further improved and
extended to build a performance profile to compare different tradeoffs.

This pull request introduces a new runtime checkpointing feature for
performance benchmarking and debugging in the RAPIDS tools. The main
changes include adding new classes for runtime checkpoints, updating
build properties, and modifying existing classes to integrate the new
checkpointing functionality.

### Memory evaluation:

- In order to enable the runtime injection set the build property
`benchmarks.checkpoints` to `dev`. This is achieved by changing the
pom.xml file or passing the property as an arhument to the mvn command.
- This property will dump the free memory after the eventlog is
processed.
- An initial test on a sample eventlog showed that it saved
approximately ~ 7MB. Of course, this resulkt is a function of the number
of completed stages in the eventlog and the accumulables associated with
each one.

### Code Changes

#### Main core changes

*
[`core/src/main/scala/org/apache/spark/sql/rapids/tool/util/stubs/StageInfoStub.scala`](diffhunk://#diff-01f45324c0321580435ceea1a21d411cc404ca921cb876ef10adb1471ff6ec88R1-R53):
Added a new class `StageInfoStub` to provide a consistent interface for
`StageInfo` across different Spark versions.
*
[`core/src/main/scala/org/apache/spark/sql/rapids/tool/store/StageModel.scala`](diffhunk://#diff-aeaf0202fd8fce5698a5c11019f3908a4692218a5ab5a47aa21796dc8884ff1cL2-R2):
Updated `StageModel` to use `StageInfoStub` for compatibility with
different Spark versions.
[[1]](diffhunk://#diff-aeaf0202fd8fce5698a5c11019f3908a4692218a5ab5a47aa21796dc8884ff1cL2-R2)
[[2]](diffhunk://#diff-aeaf0202fd8fce5698a5c11019f3908a4692218a5ab5a47aa21796dc8884ff1cR23)
[[3]](diffhunk://#diff-aeaf0202fd8fce5698a5c11019f3908a4692218a5ab5a47aa21796dc8884ff1cL34-R44)

#### Build and Configuration Updates:

*
[`core/pom.xml`](diffhunk://#diff-8d04401f1cc51365fe3e32f019cd720135ba920a1a7da7f19e9c9208478701fcR420):
Added a new property `benchmarks.checkpoints` to manage the
checkpointing feature.
*
[`core/src/main/resources/configs/build.properties`](diffhunk://#diff-4ee540ba944e31033db425f18c11f87cd3023b192fc4e84c23c56bf3ef73c95dL2-R2):
Updated build properties to include `benchmarks.checkpoints`.
[[1]](diffhunk://#diff-4ee540ba944e31033db425f18c11f87cd3023b192fc4e84c23c56bf3ef73c95dL2-R2)
[[2]](diffhunk://#diff-4ee540ba944e31033db425f18c11f87cd3023b192fc4e84c23c56bf3ef73c95dR26)

#### Runtime Checkpointing Feature:

*
[`core/src/main/scala/org/apache/spark/rapids/tool/benchmarks/DevRuntimeCheckpoint.scala`](diffhunk://#diff-19f5725d6e379e5b99a1f5f7e2d732e38a043892a772822b5b966eba86d72309R1-R37):
Added a new class `DevRuntimeCheckpoint` to insert memory markers and
print memory information during runtime for performance metrics.
*
[`core/src/main/scala/org/apache/spark/rapids/tool/benchmarks/NoOpRuntimeCheckpoint.scala`](diffhunk://#diff-9f09d79c05a71c7b8167c76f0b580fba7025ecd36a4f61edcdeeeb7fbc94bbd3R1-R29):
Added a new class `NoOpRuntimeCheckpoint` as a default no-operation
implementation for checkpoints.
*
[`core/src/main/scala/org/apache/spark/rapids/tool/benchmarks/RuntimeCheckpointTrait.scala`](diffhunk://#diff-d973ea161c3246fb40315695e67ee011b37dd43b3c912105ce4ec0ebf2512c5bR1-R29):
Introduced a new trait `RuntimeCheckpointTrait` defining the API for
inserting runtime checkpoints.
*
[`core/src/main/scala/org/apache/spark/rapids/tool/benchmarks/RuntimeInjector.scala`](diffhunk://#diff-9e52f67c406c759b6781e75e533418486b30be686214497d0512b9730b8fa027R1-R44):
Added a new object `RuntimeInjector` to manage and insert runtime
checkpoints based on build properties.


#### Integration of Checkpoints:

*
[`core/src/main/scala/org/apache/spark/sql/rapids/tool/AppBase.scala`](diffhunk://#diff-86930ce0c9a1fa003be4b6c67827f5ea8ad2c631485286d36217d9dc2bfc30e2R36):
Integrated `RuntimeInjector` to insert a memory marker after processing
events.
[[1]](diffhunk://#diff-86930ce0c9a1fa003be4b6c67827f5ea8ad2c631485286d36217d9dc2bfc30e2R36)
[[2]](diffhunk://#diff-86930ce0c9a1fa003be4b6c67827f5ea8ad2c631485286d36217d9dc2bfc30e2R496)
*
[`core/src/main/scala/org/apache/spark/sql/rapids/tool/util/RuntimeUtil.scala`](diffhunk://#diff-17ba210baa6a6a06a98ea4fc3d67a576d28f8a7e7a52be3b07c08b73f78ac4b5L2-R2):
Added a method `getJVMHeapInfo` to retrieve JVM heap information.
[[1]](diffhunk://#diff-17ba210baa6a6a06a98ea4fc3d67a576d28f8a7e7a52be3b07c08b73f78ac4b5L2-R2)
[[2]](diffhunk://#diff-17ba210baa6a6a06a98ea4fc3d67a576d28f8a7e7a52be3b07c08b73f78ac4b5R106-R117)

---------

Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>