Disable Per-SQL summary text output (#1530)
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Fixes #1527

Disable the text format output generated per-sql. The target is to
reduce the noise of the stdout and improve the performance of the
core-tools

### Impact on the output:

- `rapids_4_spark_qualification_output_persql.log` is not generated
anymore by the qualTool
- remove column `AppName` from
`rapids_4_spark_qualification_output_persql.csv`
- the `rapids_4_spark_qualification_output_persql.log` can still be
generated by the RunningQualificationApp
- The order of the SQls in the CSV file has changed. Sorted Desc based
on (GPU opportunity, and DF duration) whithin each app. Perviously, the
SQLs were sorted globally which might cause considerable overhead for a
large number of eventogs.

### Impact on Performance and usability:

- Improve readability of the stdout/log generated by the tools.
- Reduce the size of lines consumed by the python wrapper.
- Sorting Sqls per-app implies less memory requirements since it is only
required to maintain the list of SQL for the current iteration.
- Improve the string construction by avoiding filling `Buffer<String,
Int>`
- Improve the performance by skipping generating and writing the log
file to the disk.