Add comment for each expected_raw_feature indicating CSV source (#1547)
This PR adds a comment for each feature in the qualx
`expected_raw_features` list, which indicates:

- the source CSV file produced by the Profiler tool
- the source column name in parentheses if different than the feature
name, e.g. `_mean` features derived from `_sum` features.
- "N/A", if the feature is derived from other features (or other
information outside of the CSV files, e.g. `scaleFactor`).

The hope is that this list (and the comments) can be maintained as the
features evolve.

Signed-off-by: Lee Yang <leewyang@gmail.com>