Generate a detailed report for the write ops (#1544)
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Fixes #1536

This commits adds a new report containing the write operations.
The generated report is mostly functional for
`InsertIntoHadoopFsRelationCommand` and more incremental improvements
need to follow.

- New report:

No text format is generated for this report. The reason is that it is
not readable to have the path and the node description into Text and it
will represent unecessary overhead

The generated report is called `write_operations.csv` and it looks like
the following:

- `sqlPlanVersion` the number of the plan version. This is helpful if we
want to generate the write operations for all the plans when AQE is
enabled.
- `fromFinalPlan`: True when the plan is final.
- If fields are not extracted, it will insert "`unknown`"
- the `fullDescription` contains the actual node-secr truncated at 500
characters.
- `outputColumns`: lists the columns separated by semicolumn. In case of
truncated schema, the column will include something like `... and 24
more fields`.


---------

Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>