Disable dataproc enhanced optimizer configs (#1554)
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Fixes #1550

For dataproc GPU clusters, the Autotuner should set the following 2
properties to false. In addition, the autotuner appends a comment to
warn the user that those properties might be problematic to the GPU run.

- "spark.dataproc.enhanced.optimizer.enabled": "false"
- "spark.dataproc.enhanced.execution.enabled": "false"

The autotuner output will add a message stating that:

"should be disabled. WARN: Turning this property on might case the GPU
accelerated Dataproc cluster to hang."