Add sqlID column to failed_jobs.csv (#1567)
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Fixes #1563

- adds column `sqlID` to failed_jobs.csv
- the column might be empty if the job has no sqlID attached to it

This pull request includes several changes to improve the handling of
job profiling and file format extraction in the RAPIDS plugin for Apache
Spark. The most important changes include modifying the
`FailedJobsProfileResults` case class to include an optional SQL ID,
updating related views and tests, and simplifying the code in the
`HealthCheckSuite` class.