Packaging tools jar with the python package (#1634)
Fixes #1496 

This PR enables changes for packaging the tools jar with the python whl
file. This avoids downloading the latest jar from mvn each time any cli
command is run.
This packages the jar from source and bundles it in the wheel in the
folder tools_resources.

This changes the build process by packaging the jar as part of the wheel
in all build scenarios. In case of a fat build, it also packages the
csp-resources, csp-resources.tgz along with the tools jar.

The user_tools expects the tools jar now as part of the resources. In
case not available, it still falls back to downloading it from maven

The changes have been tested locally using the --platform onprem. Since
there are no platform specific changes needed, it should be sufficient.

This pull request includes changes to the `user_tools/build.sh` script
and related files to add support for a "non-fat" mode, which only
downloads necessary resources instead of all dependencies. Additionally,
it introduces a new folder structure to separate tools resources from
CSP resources.

Build script enhancements:

*
[`user_tools/build.sh`](diffhunk://#diff-33b48dfaa0c6bf1bbed25c76e81b86d6c4d3e177f220d0ebf07634c11cedea5fR32):
Added a new variable `TOOLS_RESOURCE_FOLDER` to define the tools
resources folder and updated the `download_web_dependencies` function to
handle both fat and non-fat modes.
[[1]](diffhunk://#diff-33b48dfaa0c6bf1bbed25c76e81b86d6c4d3e177f220d0ebf07634c11cedea5fR32)
[[2]](diffhunk://#diff-33b48dfaa0c6bf1bbed25c76e81b86d6c4d3e177f220d0ebf07634c11cedea5fL65-R83)
*
[`user_tools/build.sh`](diffhunk://#diff-33b48dfaa0c6bf1bbed25c76e81b86d6c4d3e177f220d0ebf07634c11cedea5fR109-R125):
Modified the `build` function to incorporate the new build modes and
ensure the appropriate resources are downloaded and organized.

Python tool context updates:

*
[`user_tools/src/spark_rapids_pytools/rapids/tool_ctxt.py`](diffhunk://#diff-478d74668cb7f8438192c9f044fe6312bd2f11df130607745e51223aa039004dR48-R50):
Introduced a new class variable `tools_resource_path` to manage tools
resources and updated methods to handle the new folder structure and
non-fat mode.
[[1]](diffhunk://#diff-478d74668cb7f8438192c9f044fe6312bd2f11df130607745e51223aa039004dR48-R50)
[[2]](diffhunk://#diff-478d74668cb7f8438192c9f044fe6312bd2f11df130607745e51223aa039004dL145-R150)
[[3]](diffhunk://#diff-478d74668cb7f8438192c9f044fe6312bd2f11df130607745e51223aa039004dL158-R181)
[[4]](diffhunk://#diff-478d74668cb7f8438192c9f044fe6312bd2f11df130607745e51223aa039004dL176-L181)
[[5]](diffhunk://#diff-478d74668cb7f8438192c9f044fe6312bd2f11df130607745e51223aa039004dL196-R210)
[[6]](diffhunk://#diff-478d74668cb7f8438192c9f044fe6312bd2f11df130607745e51223aa039004dL240-R261)

Resource manager script changes:

*
[`user_tools/src/spark_rapids_pytools/resources/dev/prepackage_mgr.py`](diffhunk://#diff-48f4c8d291d66fc725353b1a3d40b1a126a974b64c6e1daa149f3f765654fdc5L39-R40):
Added support for the `fetch_all_csp` flag to control whether all CSP
resources are downloaded or just the tools jar. Updated the script to
handle the new tools resources directory.
[[1]](diffhunk://#diff-48f4c8d291d66fc725353b1a3d40b1a126a974b64c6e1daa149f3f765654fdc5L39-R40)
[[2]](diffhunk://#diff-48f4c8d291d66fc725353b1a3d40b1a126a974b64c6e1daa149f3f765654fdc5L73-R87)
[[3]](diffhunk://#diff-48f4c8d291d66fc725353b1a3d40b1a126a974b64c6e1daa149f3f765654fdc5R101-R102)
[[4]](diffhunk://#diff-48f4c8d291d66fc725353b1a3d40b1a126a974b64c6e1daa149f3f765654fdc5L113-R123)
[[5]](diffhunk://#diff-48f4c8d291d66fc725353b1a3d40b1a126a974b64c6e1daa149f3f765654fdc5L127-R151)
[[6]](diffhunk://#diff-48f4c8d291d66fc725353b1a3d40b1a126a974b64c6e1daa149f3f765654fdc5R177-R189)

Other minor changes:

*
[`user_tools/pyproject.toml`](diffhunk://#diff-4567c711bb2a94833399095b675b18e161015821d727f84acb42052a55996616L105-R105):
Added `*.jar` to the package data to ensure the tools jar is included in
the package.

---------

Signed-off-by: Sayed Bilal Bari <sbari@nvidia.com>