Set hiveParquet and hiveORC as supported in Qual tool (#1691)
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Fixes #1588, Fixes #1671

- Enforces new hive formats in the supportedDataSources.csv
- This will mark hive format ops to be supported assuming that the users
will use built in native orc/hive support. The qual will still mark the
Read/Write as "`hive*`" to distinguish it from the native format. This
should be helpful to troubleshoot the user actually did not set the
native conversion in the GPU job
- This is an optimistic implementation since we ignore evaluating the
conditions to support those ops on the GPU.