QualX to use the pre-packaged tools_jar as part of whl (#1722)
Contributes to #1669

Currently as part of any QualX runs ( train, evaluate ), QualX expects
the environment variable `$SPARK_RAPIDS_TOOLS_JAR` to be set.
As as result of #1496 , the whl package now contains the tools_jar which
can be use directly with the user_tools commands.
We should start using the packaged jar file instead of expecting the
user to set the `SPARK_RAPIDS_TOOLS_JAR` env var

This PR uses the existing logic to find core tools jar from resources
and updates the required env variables during QualXTool run

- `SPARK_RAPIDS_TOOLS_JAR` env var already set by the user precedes the
one being set by using the resource jar
- Validation is added to check if the env var value set already actually
points to a jar
- Logs a warning in case not set and goes on to set it from the
resources
- In case no jar found in the resources, the run fails

This has been locally to run spark_rapids train command with a set of
onprem eventlogs and it completes to success

---------

Signed-off-by: sayedbilalbari <145510907+sayedbilalbari@users.noreply.github.com>