Adding QualCoreHandler to the context instead of passing to individual functions (#1745)
As part of existing patch
[PR](https://github.com/NVIDIA/spark-rapids-tools/pull/1735),
QualCoreHandler class was introduced which is concerned with all changes
related to reading QualCore output

This PR updates the existing code to use the CoreHandler from context
rather than passing it around to individual functions.
Also updates the existing definitions in the qualOutputTable for the
tables.

---------

Signed-off-by: Sayed Bilal Bari <sbari@nvidia.com>