Improve AppSqlPlanAnalyzer for large eventlogs (#1747)
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Fixes #1737

- remove redundant checks of `sqlPlanAdaptiveMetrics`
- improve the implementation of `aggregateSQLStageInfo` by changing the
dataStructure of the map.
- reduce the overhead of plan analysis by cashing the toolsGraph object
in the sqlPlan

### Impact on Performance

#### Profiling Tool

For large eventlogs, this PR has a significant impact on runtime.

A large eventlog of 6G compact size 

| property|Before | After |
|-------------|--------|--------|
|runtime| 158.8 minutes | 14.5 minutes | 
|tool| profiling | profiling |
|xmx| 24G | 24G |
|xms| 16G | 16G |

#### Qualification Tool

The analysis of the eventlog took almost the same as the profiling tool.
However, for a large eventlog, the qualification spends almost 60% of
the runtime (80 minutes) in generating the reports. The `unsupportedOps`
report is the most consuming piece of reports. Further inverstigation
can be done after merging the per-app-qual branch which includes changes
to the report generation of the qualification tool.


---------

Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>