Add modular categories to changelog release action (#1853)
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Fixes #1827

This change updates the changelog workflow and reorganizes changelog
category configuration to improve release note automation and
categorization.
We need to have better categorization of the PRs based on the priority
of the change. Previously, we had 2 main categories
`user_tools/core_tools` which does not capture the main changes in a
release.

---------

Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>