Use Tools-API to load qual reports in QualX (#1857)
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Contributes to #1836

This pull request refactors and enhances the handling of qualification
and prediction report data in the RAPIDS tools, focusing on improved
flexibility, clearer separation between per-app and global reports, and
more robust error handling and reporting. The most significant changes
are in the `CSVReportCombiner` class, which now supports both
per-application and global reports with unified processing and clearer
APIs.

- creates unit-test for the main behavioral expectations of the
result-handler
- use CSVReportCombiner and APIHelpers to load the qual data in QualX.

---------

Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>