Finalize the toolsAPI helper methods and classes (#1893)
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Fixes #1887

Simplifies the interface in order to use the Tools-API more efficiently.

This pull request refactors the API v1 report handler system in the
RAPIDS PyTools codebase to simplify and modernize the way core report
handlers are created and used. The main change is the removal of the
legacy `APIHelpers` builder pattern in favor of direct class-based
instantiation for report handlers (e.g., `QualCore`, `ProfCore`). This
results in cleaner, more readable code and a more consistent interface
for handling reports. Additionally, the `CombinedDFBuilder` is replaced
with a new `CombinedCSVBuilder` class and metaclass for combining CSV
reports.

---------

Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>