Upgrade default spark-version to 3.5.6 (#1903)
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Fixes #1601

- Upgrade default spark to 3.5.6
- remove mvn test on spark 3.4.x

We need this change because:
- 3.5.6 has bug fixes. Especially, the json string limitation of
https://issues.apache.org/jira/browse/SPARK-49872
- reduce the fat-wheel since we will need only 1 version of spark
(3.5.6)
- reduce build and testing between on-prem/csps as they have same spark
release by default.

Pairing Dependencies:

- The search results from Maven Repository and Scaladex show that
`spark-hadoop-cloud_2.13-3.5.6` and `spark-hadoop-cloud_2.12-3.5.6` both
depend on `gcs-connector:hadoop3-2.2.14`. Therefore, the compatible GCS
connector is `gcs-connector-hadoop3-2.2.14-shaded.jar`.
- The version of Hadoop's AWS module and the AWS SDK are part of the
specific dependency set tested and bundled with a given Spark release.
- `hadoop-aws-3.4.1.jar: `Spark 3.5.6 officially depends on Hadoop
3.4.1. The hadoop-aws jar is a component of Hadoop itself, so it needs
to be compatible with the Hadoop version used by Spark.
- `aws-java-sdk-bundle-1.12.262.jar`: version 1.12.262 is the specific
version that the hadoop-aws-3.4.1 library has been built and tested
with.
- The compatible hadoop-azure jar is `hadoop-azure-3.3.4.jar`. This is
the same version used with against Spark 3.5.0. This is because the
spark-hadoop-cloud module for Spark 3.5.6, which provides the connectors
for cloud storage, still lists hadoop-azure version 3.3.4 as a
dependency.

---------

Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>