Add Iceberg to core pom dependency
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Contributes to #1919

This change is to add support to run/generate tests with IceBerg. Prior
to that, core-tools did not handle IceBerg.
For this initial change, the `AppenData` is considered unsupported.
In followups to the same issue, we need to write Parser for AppenData
and be able to recongnize it as a write Op along with the approprate
Format.
We might need to define a separate Helper for Iceberg (similar to
DeltaLake/Databricks)