Use hadoop-aws 3.3.4 with spark-356 and 357 (#1963)
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

contributes to #1962

Hadoop-aws 3.4 is not compatible with aws-sdk 1.x
This code forces aws back to 3.3.4

Hadoop-aws 3.4 causes an exception. this exception disappears when
switching to 3.3.4
```
        | Exception in thread "main" java.lang.NoClassDefFoundError: software/amazon/awssdk/core/exception/SdkException
        |       at java.base/java.lang.Class.forName0(Native Method)
        |       at java.base/java.lang.Class.forName(Class.java:467)
        |       at org.apache.hadoop.conf.Configuration.getClassByNameOrNull(Configuration.java:2625)
```

The original issue is planned to change baws-sdk jar. However, this is a
big change because the entire file would require changes in the
enviornment variables and how the default behavior works.