Disable text output format by default (#1982)
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Fixes #1793

- disable the text format reports of large tables. In legacy output, the
`profile.log` could grow up to tens of MB. The generation of text format
of large files puts more stress on the heap.
- the change saves 12% of execution time and reduces the memory usage
significantly for each app.
- Addds a "plugin" pattern to allow extending the appProperties analysis
in a more standard way. This helps in future work to support more
features in a plugin fashion.

This pull request introduces a new plugin-based framework for detecting
and managing Spark application features (such as Auron and Delta Lake
OSS) via application properties, and refactors the detection logic to
use this extensible system. It also adds support for the new Auron
runtime and improves version extraction for plugins. The changes are
grouped below by theme.