Use plugin-based approach to process eventlogs (#2008)
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Fixes #2007, Contributes to #1978

This commit introduces significant improvements to the plugin system and
reorganizes code for better maintainability and logical grouping.
This pull request refactors and improves the detection and handling of
Databricks and Photon runtimes in the codebase. The main focus is on
modularizing Databricks-specific logic into a dedicated `db` package,
introducing a new `DBPlugin` class for runtime detection, and updating
references throughout the code to use the new structure. Additionally,
some minor code and documentation improvements are included.